Improves enemy and player networking
Adds NetworkAnimator and NetworkTransform components to player and enemy prefabs.

Enables network synchronization of animations and transforms for smoother multiplayer gameplay.

Fixes the problem of enemies getting disabled due to a Mirror issue

The NavMeshAgent is disabled on the client side to make sure that the movement on screen is what is seen server side

Adds ability to force quit